Move DualizeConeProg into affine2direct and clean up dead code
- Move DualizeConeProg from its own file into affine2direct.py
  (it's the only consumer of Dualize.invert, not a cone2cone)
- Delete dualize_cone_prog.py
- Remove svec_to_full_mat (dead: PSD goes through dualized path,
  never hits ACC extract_dual_value)
- Remove PSD branch from extract_dual_value (same reason)
- Simplify solving chain: use reductions.append directly
