+++ cvxpy/reductions/cone2cone/affine2direct.py
@@ -24,6 +24,8 @@ from cvxpy.constraints.power import PowCone3D as PowCone_obj
 from cvxpy.constraints.psd import PSD as PSD_obj
 from cvxpy.constraints.second_order import SOC as SOC_obj
 from cvxpy.constraints.zero import Zero as Zero_obj
+from cvxpy.reductions.dcp2cone.cone_matrix_stuffing import ParamConeProg
+from cvxpy.reductions.reduction import Reduction
 from cvxpy.reductions.solution import Solution
 
 FREE = 'fr'
@@ -429,3 +431,64 @@ class Slacks:
         if solution.opt_val is not None:
             solution.opt_val += inv_data[s.OBJ_OFFSET]
         return solution
+
+
+class DualizeConeProg(Reduction):
+    """Flags a ParamConeProg for dualized interpretation by the solver.
+
+    This reduction sits between ConeMatrixStuffing and the solver.  It
+    does **not** rewrite any matrices — it simply sets the ``dualized``
+    flag on the ``ParamConeProg`` so that downstream solvers interpret
+    the existing data ``(A, b, c, d, P, cone_dims)`` as the Lagrangian
+    dual:
+
+        (D)  max  -b'y  [- ½w'Pw]  + d
+             s.t. A'y   [+ Pw]      = c
+                  y in K*
+
+    where ``K*`` is the dual of the cone ``K`` described by
+    ``cone_dims``.  The slack variable ``w`` (size ``x.size``) is
+    introduced by the solver only when ``P`` is not ``None``.
+
+    Primal recovery
+    ---------------
+    ``x*`` is obtained from the dual of the equality constraint
+    ``A'y = c`` (equivalently, ``A'y + Pw = c`` when ``P`` is
+    present).  Original constraint duals are the optimal ``y``
+    blocks.
+
+    The solver is responsible for:
+    1. Interpreting the flag (building a max problem, placing cones
+       on variables, etc.).
+    2. Returning a ``Solution`` whose ``primal_vars`` and
+       ``dual_vars`` follow the ``Dualize`` convention so that
+       ``invert`` can recover the original solution.
+
+    Position in the chain
+    ---------------------
+    ::
+
+        ... -> ConeMatrixStuffing -> DualizeConeProg -> Solver
+    """
+
+    def accepts(self, problem) -> bool:
+        return (isinstance(problem, ParamConeProg)
+                and not problem.is_mixed_integer()
+                and not problem.dualized)
+
+    def apply(self, problem):
+        problem.dualized = True
+        inv_data = {
+            'x_id': problem.x.id,
+            'constr_map': problem.constr_map,
+        }
+        return problem, inv_data
+
+    def invert(self, solution, inverse_data):
+        """Delegates to Dualize.invert for solution recovery."""
+        inv = {
+            s.OBJ_OFFSET: 0,  # offset already baked into solver's opt_val
+            'constr_map': inverse_data['constr_map'],
+            'x_id': inverse_data['x_id'],
+        }
+        return Dualize.invert(solution, inv)
