+++ cvxpy/reductions/solvers/solving_chain.py
@@ -18,8 +18,8 @@ if TYPE_CHECKING:
     from cvxpy.problems.problem import Problem
 from cvxpy.reductions.chain import Chain
 from cvxpy.reductions.complex2real import complex2real
+from cvxpy.reductions.cone2cone.affine2direct import DualizeConeProg
 from cvxpy.reductions.cone2cone.approx import ApproxCone2Cone
-from cvxpy.reductions.cone2cone.dualize_cone_prog import DualizeConeProg
 from cvxpy.reductions.cone2cone.exact import ExactCone2Cone
 from cvxpy.reductions.cone2cone.soc_dim3 import SOCDim3
 from cvxpy.reductions.cvx_attr2constr import CvxAttr2Constr
@@ -227,13 +227,11 @@ def _build_solving_chain(
     if solver_instance.SOC_DIM3_ONLY and SOC in cones:
         reductions.append(SOCDim3())
 
-    stuffing_and_solver = [
-        ConeMatrixStuffing(quad_obj=quad_obj, canon_backend=canon_backend),
-    ]
+    reductions.append(
+        ConeMatrixStuffing(quad_obj=quad_obj, canon_backend=canon_backend))
     if solver_instance.should_dualize(problem_form):
-        stuffing_and_solver.append(DualizeConeProg())
-    stuffing_and_solver.append(solver_instance)
-    reductions += stuffing_and_solver
+        reductions.append(DualizeConeProg())
+    reductions.append(solver_instance)
     return SolvingChain(reductions=reductions, solver_context=solver_context)
